Update App.js
Removed the if else for the login page as a quick fix, to resolve this issue correctly, you would need to add an if else for a user is null I think in the api. This allows you to access the login page.
Shouldn't matter because I'll be rolling back and deleting or commenting out all login features.